base64url_encode {jose}	R Documentation

Base64URL encoding

Description

The base64url_encode functions are a variant of the standard base64. They are specified in Section 5 of RFC 4648 as a URL-safe alternative. They use different symbols for the 62:nd and 63:rd alphabet character and do not include trailing == padding.

Usage

base64url_encode(bin)

base64url_decode(text)

Arguments

bin
a binary blob to encode

text
a base64url encoded string

[Package jose version 1.2.1 Index]